refactor: place template workspace hosts under presentation
Move coordinator, binders, sort dialog, and presentation snapshot into templates.presentation so JVM coverage follows the directory contract.
