+++ app/src/main/java/com/dpis/module/templates/QuickTemplateTargetCatalogLoader.kt
@@ -4,6 +4,7 @@ import com.dpis.module.DpisLog
 import com.dpis.module.applist.InstalledAppCatalogCoordinator
 import com.dpis.module.applist.InstalledAppCatalogItem
 import com.dpis.module.config.PackageConfigRepository
+import com.dpis.module.templates.presentation.QuickTemplateTargetsBinder
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
