+++ app/src/main/java/com/dpis/module/templates/TemplateDetailPaneController.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.FrameLayout
 import androidx.core.view.ViewCompat
 import com.dpis.module.R
+import com.dpis.module.templates.presentation.QuickTemplateTargetsBinder
 import com.dpis.module.ui.WindowInsetsBinder
 
 /** Owns the legacy landscape template-detail view and its target-picker lifetime. */
