+++ app/src/test/java/com/dpis/module/templates/QuickTemplateApplySourceSmokeTest.java
@@ -11,8 +11,8 @@ public class QuickTemplateApplySourceSmokeTest {
     @Test
     public void templateWorkspaceCoordinatorOwnsApplyConfirmationAndResultCopy() throws IOException {
         String mainActivity = read("src/main/java/com/dpis/module/MainActivity.java");
-        String workspace = read("src/main/java/com/dpis/module/templates/TemplateWorkspaceCoordinator.kt");
-        String binder = read("src/main/java/com/dpis/module/templates/TemplateWorkspaceBinder.kt");
+        String workspace = read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceCoordinator.kt");
+        String binder = read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceBinder.kt");
         String coordinator = read(
                 "src/main/java/com/dpis/module/templates/QuickTemplateApplyCoordinator.kt");
         String adapters = read("src/main/java/com/dpis/module/templates/QuickTemplateApplyAdapters.java");
