+++ app/src/test/java/com/dpis/module/templates/TemplateWorkspaceLayoutSmokeTest.kt
@@ -47,10 +47,10 @@ class TemplateWorkspaceLayoutSmokeTest {
         )
         card.assertNotContainsAll("android:id=\"@+id/quick_template_updated\"", "android:id=\"@+id/quick_template_missing_font\"")
         assertDashedEmptySummaryState(elementWithId(card, "quick_template_empty_summary"))
-        read("src/main/java/com/dpis/module/templates/QuickTemplateSortDialog.kt").assertContainsAll("QuickTemplateSortContent(", "ReorderableItem", "longPressDraggableHandle", "R.drawable.ic_drag_indicator_24")
+        read("src/main/java/com/dpis/module/templates/presentation/QuickTemplateSortDialog.kt").assertContainsAll("QuickTemplateSortContent(", "ReorderableItem", "longPressDraggableHandle", "R.drawable.ic_drag_indicator_24")
         read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceContent.kt").assertContainsAll("var sortDialogVisible by rememberSaveable", "QuickTemplateSortDialog(", "items = state.sortItems", "onOrderChanged = state.actions::reorderTemplates")
-        read("src/main/java/com/dpis/module/templates/TemplateWorkspacePresentation.kt").assertContainsAll("val sortItems: List<QuickTemplateSortItem>", "fun reorderTemplates(orderedIds: List<String>): Boolean")
-        read("src/main/java/com/dpis/module/templates/TemplateWorkspaceCoordinator.kt").assertContainsAll(
+        read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspacePresentation.kt").assertContainsAll("val sortItems: List<QuickTemplateSortItem>", "fun reorderTemplates(orderedIds: List<String>): Boolean")
+        read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceCoordinator.kt").assertContainsAll(
             "class TemplateWorkspaceCoordinator", "private val presentation = TemplateWorkspacePresentationController",
             "refresh(presentation.state().query)",
             "QuickTemplateStore(activity).reorder(orderedIds)",
@@ -63,7 +63,7 @@ class TemplateWorkspaceLayoutSmokeTest {
 
     @Test
     fun binderAndAdapterReadStoresAndBindMissingFontHooks() {
-        val binder = read("src/main/java/com/dpis/module/templates/TemplateWorkspaceBinder.kt")
+        val binder = read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceBinder.kt")
         val adapter = read("src/main/java/com/dpis/module/templates/QuickTemplateListAdapter.java")
         binder.assertContainsAll("GlobalPrefillStore(preferences).read()", "QuickTemplateStore(context).readAll()", "ConfigStoreFactory.createLocalUiFontLibraryStore", "TemplateTypefaceResolver(", "store.resolveFontFile(typefaceId) != null", "R.id.global_prefill_summary_chips", "bindHeaderActions(workspaceView, templates)", "sortButton.isEnabled = enabled", "sortButton.alpha = if (enabled) 1f else TemplateUiTokens.DISABLED_ACTION_ALPHA")
         read("src/main/java/com/dpis/module/templates/TemplateTypefaceResolver.kt").assertContainsAll(
@@ -73,8 +73,8 @@ class TemplateWorkspaceLayoutSmokeTest {
         read("src/main/java/com/dpis/module/templates/TemplateSummaryChipBinder.java").assertContainsAll("colorSurfaceContainerHighest", "chip.setChipStrokeWidth(0);")
         adapter.assertContainsAll("R.id.quick_template_summary_chips", "TemplateSummaryChipBinder", "R.id.quick_template_apply_button", "R.id.quick_template_edit_button", "R.id.quick_template_select_button")
         adapter.assertNotContainsAll("quick_template_updated")
-        read("src/main/java/com/dpis/module/templates/TemplateWorkspaceBinder.kt").assertContainsAll("quick_template_sort_button")
-        read("src/main/java/com/dpis/module/templates/QuickTemplateSortDialog.kt").assertContainsAll("ModalDialog(onDismissRequest = onDismiss)", "fun QuickTemplateSortDialog(")
+        read("src/main/java/com/dpis/module/templates/presentation/TemplateWorkspaceBinder.kt").assertContainsAll("quick_template_sort_button")
+        read("src/main/java/com/dpis/module/templates/presentation/QuickTemplateSortDialog.kt").assertContainsAll("ModalDialog(onDismissRequest = onDismiss)", "fun QuickTemplateSortDialog(")
         read("src/main/java/com/dpis/module/ui/presentation/wear/WearWorkspaceContent.kt").assertContainsAll("var sortDialogVisible by rememberSaveable", "QuickTemplateSortDialog(", "enabled = state.sortItems.isNotEmpty()")
         read("src/main/java/com/dpis/module/MainActivity.java").apply {
             assertContainsAll("private TemplateWorkspaceActivitySession workspaceSession;", "ensureWorkspaceSession()")
