+++ sonar-project.properties
@@ -32,11 +32,6 @@ app/src/main/java/com/dpis/module/diagnostics/ForegroundAppReader.kt,\
 app/src/main/java/com/dpis/module/settings/SystemScopeCoordinator.kt,\
 app/src/main/java/com/dpis/module/settings/SystemFontScaleToolBinder.java,\
 app/src/main/java/com/dpis/module/settings/ToolsWorkspaceBinder.java,\
-app/src/main/java/com/dpis/module/templates/QuickTemplateSortDialog.kt,\
-app/src/main/java/com/dpis/module/templates/TemplateWorkspaceCoordinator.kt,\
-app/src/main/java/com/dpis/module/templates/TemplateWorkspacePresentation.kt,\
-app/src/main/java/com/dpis/module/templates/TemplateWorkspaceBinder.kt,\
-app/src/main/java/com/dpis/module/templates/QuickTemplateTargetsBinder.kt,\
 app/src/main/java/com/dpis/module/updates/UpdatePackageInstaller.java,\
 app/src/main/java/com/dpis/module/updates/UpdateAvailableDialog.kt,\
 app/src/main/java/com/dpis/module/updates/ReleaseNotesMarkwonPlugin.kt,\
